+++ Shared/Coordinators/SettingsCoordinator.swift
@@ -79,10 +79,8 @@ final class SettingsCoordinator: NavigationCoordinatable {
     var videoPlayerSettings = makeVideoPlayerSettings
     @Route(.modal)
     var playbackQualitySettings = makePlaybackQualitySettings
-    @Route(.push)
+    @Route(.modal)
     var userProfile = makeUserProfileSettings
-    @Route(.push)
-    var localSecurity = makeLocalSecurity
     #endif
 
     #if os(iOS)
@@ -185,23 +183,24 @@ final class SettingsCoordinator: NavigationCoordinatable {
     #endif
 
     #if os(tvOS)
-    @ViewBuilder
-    func makeUserProfileSettings(viewModel: SettingsViewModel) -> some View {
-        UserProfileSettingsView(viewModel: viewModel)
-    }
 
-    @ViewBuilder
-    func makeLocalSecurity() -> some View {
-        UserLocalSecurityView()
-    }
+    // MARK: - User Profile View
 
-    func makeCustomizeViewsSettings() -> NavigationViewCoordinator<BasicNavigationViewCoordinator> {
+    func makeUserProfileSettings(viewModel: SettingsViewModel) -> NavigationViewCoordinator<UserProfileSettingsCoordinator> {
         NavigationViewCoordinator(
-            BasicNavigationViewCoordinator {
-                CustomizeViewsSettings()
-            }
+            UserProfileSettingsCoordinator(viewModel: viewModel)
         )
     }
+
+    // MARK: - Customize Settings View
+
+    func makeCustomizeViewsSettings() -> NavigationViewCoordinator<CustomizeSettingsCoordinator> {
+        NavigationViewCoordinator(
+            CustomizeSettingsCoordinator()
+        )
+    }
+
+    // MARK: - Experimental Settings View
 
     func makeExperimentalSettings() -> NavigationViewCoordinator<BasicNavigationViewCoordinator> {
         NavigationViewCoordinator(
@@ -211,11 +210,15 @@ final class SettingsCoordinator: NavigationCoordinatable {
         )
     }
 
+    // MARK: - Poster Indicator Settings View
+
     func makeIndicatorSettings() -> NavigationViewCoordinator<BasicNavigationViewCoordinator> {
         NavigationViewCoordinator {
             IndicatorSettingsView()
         }
     }
+
+    // MARK: - Server Settings View
 
     func makeServerDetail(server: ServerState) -> NavigationViewCoordinator<BasicNavigationViewCoordinator> {
         NavigationViewCoordinator {
@@ -223,11 +226,15 @@ final class SettingsCoordinator: NavigationCoordinatable {
         }
     }
 
+    // MARK: - Video Player Settings View
+
     func makeVideoPlayerSettings() -> NavigationViewCoordinator<VideoPlayerSettingsCoordinator> {
         NavigationViewCoordinator(
             VideoPlayerSettingsCoordinator()
         )
     }
+
+    // MARK: - Playback Settings View
 
     func makePlaybackQualitySettings() -> NavigationViewCoordinator<PlaybackQualitySettingsCoordinator> {
         NavigationViewCoordinator(
